fixed resources
How all resources share a base class. When a resource is required it is loaded explicitly rather than as part of object instantiation, and stored in a list of resources. If a resource is requested that alrady exists in the list it provides that resource rather than creating a new one